[rs proto] 06 - Refactors component locations in ChartView directory; Adds ChartView component; moves RsPane, Chart and ChartTwo components into ChartView directory; adds ChartToolbar component; Moves utils files to utils directory;
