+++ src/app/features/rs/comps/chart-view/chart-toolbar/chart-toolbar.component.scss
@@ -1,0 +1,31 @@
+@use '../../../../../styles/mixins-global' as *;
+
+.zoom-toolbar-center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+}
+
+.zoom-controls {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: rem(16);
+  margin-bottom: rem(16);
+}
+
+.zoom-status {
+  margin-left: rem(16);
+  font-size: rem(14);
+  color: var(--theme-on-surface, #222);
+}
+
+button {
+  @include rsButtonOne();
+  font-size: rem(16);
+  padding: rem(8) rem(14);
+  margin: 0;
+  min-width: rem(36);
+  min-height: rem(36);
+}
