+++ src/app/features/rs/comps/chart-view/chart-toolbar/chart-toolbar.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ChartToolbarComponent } from './chart-toolbar.component';
+
+describe('ChartToolbarComponent', () => {
+  let component: ChartToolbarComponent;
+  let fixture: ComponentFixture<ChartToolbarComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ChartToolbarComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(ChartToolbarComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
