+++ src/app/features/rs/comps/chart-view/chart-view.component.html
@@ -1,0 +1,49 @@
+<section class="candlestick-chart-two">
+  <h2>MSFT Candlestick Chart (Syncfusion Demo)</h2>
+  <rs-chart-toolbar
+    [zoomFactor]="zoomFactor"
+    [zoomPosition]="zoomPosition"
+    (zoomIn)="zoomIn()"
+    (zoomOut)="zoomOut()"
+    (resetZoom)="resetZoom()"
+    (useSubset)="useSubset()"
+  ></rs-chart-toolbar>
+  <div class="chart-two-root">
+    <div class="chart-two-main-flex-col">
+      <div class="chart-two-plot-area" style="position: relative;">
+        <rs-pane
+          [candleData]="candleData()"
+          [plotAreaLeft]="plotAreaLeft"
+          [plotAreaWidth]="plotAreaWidth"
+          [xAxisTicks]="visibleXAxisTicks()"
+          aria-label="RS Color Pane"
+        ></rs-pane>
+        @if (hasCandleData && chartPrimaryXAxis && candleData().length > 0) {
+          <ejs-chart #msftChart id="candlestick-chart-two" [primaryXAxis]="chartPrimaryXAxis" [primaryYAxis]="{ title: 'Price (USD)', labelFormat: '${value}' }" [tooltip]="{ enable: tooltipEnabled }" (axisLabelRender)="onAxisLabelRender($event)" (tooltipRender)="onTooltipRender($event)" [zoomSettings]="{ enableSelectionZooming: true, enablePinchZooming: true, mode: 'X', enableToolbar: false }" (zoomComplete)="onChartZoomComplete($event)" (loaded)="onChartLoaded()">
+            <e-series-collection>
+              <e-series
+                [dataSource]="candleData() || []"
+                type='Candle'
+                xName='x'
+                high='high'
+                low='low'
+                open='open'
+                close='close'
+                name='MSFT'
+                pointColorMapping='rsColor'>
+              </e-series>
+            </e-series-collection>
+          </ejs-chart>
+        }
+      </div>
+      @if (firstVisibleYear() !== null) {
+        <div class="x-axis-year-label-row">
+          <span class="x-axis-year-label">{{ firstVisibleYear() }}</span>
+        </div>
+      }
+      <div class="rs-comparison-summary">
+        RS Comparison: {{ rsComparisonSummary() }}
+      </div>
+    </div>
+  </div>
+</section>
