+++ src/app/features/rs/comps/chart-view/chart-view.component.scss
@@ -1,0 +1,10 @@
+@use '../../../../styles/variables-global' as *;
+@use '../../../../styles/mixins-global' as *;
+
+.chart-view-root {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  min-height: rem(100);
+  padding: rem(16);
+}
