+++ src/app/features/rs/comps/chart-view/chart/chart.component.scss
