+++ src/app/features/rs/comps/chart-view/rs-pane/rs-pane.component.html
