+++ src/app/features/rs/comps/chart-view/rs-pane/rs-pane.component.scss
@@ -1,5 +1,5 @@
-@use '../../../../styles/variables-global' as *;
-@use '../../../../styles/mixins-global' as *;
+@use '../../../../../styles/variables-global' as *;
+@use '../../../../../styles/mixins-global' as *;
 
 .rs-pane {
   position: relative;
